week4/mqsubws/requirements.txt
pyzmq==26.2.1
paho-mqtt==2.1.0
fastapi==0.115.12
uvicorn==0.34.0
